% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_data.R
\name{df_to_tensor}
\alias{df_to_tensor}
\title{Prepare data for tensorsem model}
\usage{
df_to_tensor(df, dtype = NULL, device = NULL)
}
\arguments{
\item{df}{data frame}

\item{dtype}{data type of the resulting tensor}

\item{device}{device to store the resulting tensor on}
}
\value{
Torch tensor of scaled and processed data
}
\description{
This function prepares a dataframe for a tensorsem model. It
first converts the variables to a design matrix, then centers
it, and lastly converts it to a torch_tensor
}
\seealso{
\code{\link[torch:torch_tensor]{torch::torch_tensor()}}, \code{\link[stats:model.matrix]{stats::model.matrix()}}
}
