% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_parse.R
\name{lav_mod_to_torch_opts}
\alias{lav_mod_to_torch_opts}
\title{Create a torch options list from a lavaan Model class.}
\usage{
lav_mod_to_torch_opts(lav_mod)
}
\arguments{
\item{lav_mod}{lavaan Model class object}
}
\value{
list of tensorsem options
}
\description{
Create a torch options list from a lavaan Model class.
}
