% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lav_to_tf_pars.R
\name{lav_to_tf_pars}
\alias{lav_to_tf_pars}
\title{Convert a lavaan model to sem_tf parameters}
\usage{
lav_to_tf_pars(mod, data)
}
\arguments{
\item{mod}{lavaan model syntax}

\item{data}{data frame}
}
\description{
Convert a lavaan model to sem_tf parameters
}
\keyword{internal}
