% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loss_functions.R
\name{mvn_negloglik}
\alias{mvn_negloglik}
\title{Multivariate normal negative log-likelihood loss function for tensorsem nn module.}
\usage{
mvn_negloglik(dat, Sigma)
}
\arguments{
\item{dat}{The centered dataset as a tensor}

\item{Sigma}{The model() implied covariance matrix}
}
\value{
torch_tensor: scalar negative log likelihood
}
\description{
Multivariate normal negative log-likelihood loss function for tensorsem nn module.
}
