% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_parse.R
\name{partable_from_torch}
\alias{partable_from_torch}
\title{Create a lavaan parameter table from torch free_params output}
\usage{
partable_from_torch(pars, syntax)
}
\arguments{
\item{pars}{data frame of parameter estimates (est) and their standard errors (se)}

\item{syntax}{syntax of the original model}
}
\description{
See examples in tensorsem for how to save the output.
}
