% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loss_functions.R
\name{sem_fitfun}
\alias{sem_fitfun}
\title{SEM fitting function}
\usage{
sem_fitfun(S, Sigma)
}
\arguments{
\item{S}{The observed covariance matrix}

\item{Sigma}{The model implied covariance matrix}
}
\value{
torch_tensor: scalar loss function
}
\description{
SEM fitting function
}
