% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_parse.R
\name{syntax_to_torch_opts}
\alias{syntax_to_torch_opts}
\title{Create a torch options list from lavaan syntax.}
\usage{
syntax_to_torch_opts(syntax)
}
\arguments{
\item{syntax}{lavaan syntax}
}
\value{
list of tensorsem options
}
\description{
Create a torch options list from lavaan syntax.
}
