% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tf_dataset.R
\docType{data}
\name{tf_data}
\alias{tf_data}
\title{create tensorflow dataset}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
tf_data
}
\description{
create tensorflow dataset
}
\keyword{internal}
