% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lav_to_tf_pars.R
\name{tf_pars_to_glist}
\alias{tf_pars_to_glist}
\title{Convert tf params to lavaan GLIST}
\usage{
tf_pars_to_glist(tf_pars, type = "start")
}
\description{
Convert tf params to lavaan GLIST
}
\keyword{internal}
