% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tf_pars_to_session.R
\name{tf_pars_to_session}
\alias{tf_pars_to_session}
\title{Initialise a tensorflow environment from tf params}
\usage{
tf_pars_to_session(params)
}
\description{
Initialise a tensorflow environment from tf params
}
\keyword{internal}
