% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tf_sem_object.R
\name{tf_sem_object}
\alias{tf_sem_object}
\title{TensorFlow SEM object}
\description{
An R6 object with an initialized TensorFlow environment and user-friendly methods and bindings to this environment
}
\keyword{internal}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{tf_sem_object$new()}}
\item \href{#method-train}{\code{tf_sem_object$train()}}
\item \href{#method-print}{\code{tf_sem_object$print()}}
\item \href{#method-summary}{\code{tf_sem_object$summary()}}
\item \href{#method-gradients}{\code{tf_sem_object$gradients()}}
\item \href{#method-plot_loss}{\code{tf_sem_object$plot_loss()}}
\item \href{#method-clone}{\code{tf_sem_object$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{tf_sem_object$new(tf_session, mod, sample_size)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-train"></a>}}
\subsection{Method \code{train()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{tf_sem_object$train(niter = 200, pb = TRUE, verbose = FALSE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{tf_sem_object$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-summary"></a>}}
\subsection{Method \code{summary()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{tf_sem_object$summary()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-gradients"></a>}}
\subsection{Method \code{gradients()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{tf_sem_object$gradients()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plot_loss"></a>}}
\subsection{Method \code{plot_loss()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{tf_sem_object$plot_loss()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{tf_sem_object$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
