% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_parse.R
\name{torch_opts_to_file}
\alias{torch_opts_to_file}
\title{Export a torch options list to a file to be read by tensorsem}
\usage{
torch_opts_to_file(opts, filename = "opts.pkl")
}
\arguments{
\item{opts}{options list}

\item{filename}{path to save the list to}
}
\description{
Use the .pkl (pickle) extension
}
