% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{torch_vech}
\alias{torch_vech}
\title{Half-vectorization of square matrices}
\usage{
torch_vech(x)
}
\arguments{
\item{x}{square (symmetric) matrix tensor}
}
\value{
column vector of stacked lower-diagonal elements
}
\description{
Half-vectorization of square matrices
}
