% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tf_sem_object.R
\docType{data}
\name{tf_sem_object}
\alias{tf_sem_object}
\title{TensorFlow SEM object}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
tf_sem_object
}
\description{
An R6 object with an initialized TensorFlow environment and user-friendly methods and bindings to this environment
}
\keyword{internal}
